DEV.ENH: Add class `ProptyDevice` to handle set of PVs with same `device_name` and a commom property prefix.
