+++ siriuspy/siriuspy/devices/device.py
@@ -167,6 +167,24 @@ class Device:
         return pvname
 
 
+class ProptyDevice(Device):
+    """Device with a prefix property name."""
+
+    def __init__(self, devname, propty_prefix, properties):
+        """."""
+        self._propty_prefix = propty_prefix
+        # call base class constructor
+        super().__init__(devname, properties=properties)
+
+    def _get_pvname(self, devname, propty):
+        if devname:
+            func = devname.substitute
+            pvname = func(propty=self._propty_prefix + propty)
+        else:
+            pvname = self._propty_prefix + propty
+        return pvname
+
+
 # NOTE: This class is temporary. It should become deprecated once all
 # devices names are in accordance with Sirius naming system
 class DeviceNC(Device):
